Some changes in Add to card and checkout process

+++ ASGlass/ASGlass/Controllers/CheckoutController.cs
@@ -64,13 +64,48 @@ namespace ASGlass.Controllers
                             orders.OrderNumber = random.Next(10000,90000);
                         }
 
-                        Product product = _context.Products.FirstOrDefault(x => x.Id == item.ProductId);
+                        if (item.ProductId != null)
+                        {
+                            Product product = _context.Products.FirstOrDefault(x => x.Id == item.ProductId);
+                            _context.Products.Remove(product);
+
+                        }
 
                         _context.Orders.Add(orders);
-                        _context.Products.Remove(product);
                     }
                 }
                 HttpContext.Response.Cookies.Delete("Products");
+            }
+            else
+            {
+                List<CartItem> cartItems = _context.CartItems.Include(x => x.Product.ProductImages).Include(x => x.Product).Include(x => x.Product.Shape).Include(x => x.Product.Colors).Where(x => x.AppUserId == member.Id).ToList();
+
+                foreach (var item in cartItems)
+                {
+                    Order orders = new Order();
+                    Random random = new Random();
+                    if (cartItems != null)
+                    {
+                        orders.AppUserId = item.AppUserId;
+                        orders.ProductId = item.ProductId;
+                        orders.FullName = HttpContext.Request.Form["fullname"].ToString();
+                        orders.Email = HttpContext.Request.Form["email"].ToString();
+                        orders.ProductName = item.Name;
+                        orders.Price = item.Price;
+                        orders.ProductImage = item.Image;
+                        orders.CreatedAt = DateTime.UtcNow;
+                        orders.Status = Models.Enums.OrderStatus.Pending;
+                        orders.OrderNumber = random.Next(10000, 90000);
+                    }
+
+                    if (item.ProductId != null)
+                    {
+                        Product product = _context.Products.FirstOrDefault(x => x.Id == item.ProductId);
+                        _context.Products.Remove(product);
+
+                    }
+                    _context.Orders.Add(orders);
+                }
             }
